msm: clock: Use empty strings for devices names of unused clocks
NULL pointers in the place of device names act as wildcards that
can cause inadvertent clk_get() matches when multiple clocks in
the table have the same name. clk_get() first looks for a clock
with both a matching name and device. If none are found, it looks
for a clock where just the name matches and the device NULL.

Instead of using a NULL string in the tables for unused clocks,
specify an empty string. This will cause clk_get() to fail and
return an error if an exact match is found, rather than returning
a pointer to a wrong clock.

Change-Id: Ib3793cd52a71bb6b08b88b03dad615165596698b
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>